requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
sse-starlette==2.1.3

# AI / LangGraph / LangChain / Google ADK
langchain==0.3.14
langchain-core==0.3.28
langchain-google-genai==2.0.7
langgraph==0.2.69
google-generativeai==0.8.3
google-adk==0.4.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Config & validation
pydantic==2.10.3
pydantic-settings==2.6.1

# HTTP client
httpx==0.28.1

# Utilities
python-dateutil==2.9.0
numpy==2.2.0

# Image generation (Gemini Imagen 3 / "Nano Banana" per Google AI docs)
google-genai>=1.0.0
Pillow>=10.0.0
